integracao_supabase_importer/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pypdf==4.3.1
